wisdom-campus-backend/requirements.txt
# 智慧校园学习助手系统后端依赖
# 核心框架
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1

# 数据库
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
PyMySQL==1.0.2
cryptography==38.0.4
alembic==1.7.4
Flask-Migrate==3.1.0

# API和序列化
Flask-RESTful==0.3.9
marshmallow==3.14.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0

# 认证和安全
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==0.7.1
python-jose==3.3.0
python-dotenv==0.19.1
Authlib==1.0.1
requests==2.26.0

# 跨域支持
Flask-Cors==3.0.10

# 文档生成
Swagger-UI-py==21.12.8
PyYAML==6.0

# 测试工具
pytest==6.2.5
pytest-flask==1.2.0

# 开发工具
python-dotenv==0.19.1
black==21.9b0
flake8==4.0.1

# 工具库
Pillow==9.0.0
pydantic==1.9.0
